Fix #53 - Empêche de supprimer automatiquement les mp non lus, ainsi que les mp envoyés a des membres disposant des droits d'administrateur ou de modérateur global

+++ main/process/process_prune_pm.php
@@ -16,7 +16,8 @@ function prune_pm()
 {
 	$sql = 'DELETE FROM ' . SQL_PREFIX . 'mp
 			WHERE mp_time < ' . (CURRENT_TIME - (ONE_MONTH * 6)) . '
-				AND (mp_type = ' . MP_INBOX . ' OR mp_type = ' . MP_OUTBOX . ')';
+				AND mp_read = 1 AND (mp_type = ' . MP_INBOX . ' OR mp_type = ' . MP_OUTBOX . ')
+				AND mp_to NOT IN(SELECT u_id FROM ' . SQL_PREFIX . 'users WHERE u_auth > ' . MODO . ')';
 	Fsb::$db->query($sql);
 }
 /* EOF */
